% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRR.r
\name{atamult}
\alias{atamult}
\title{Jacobian matrix multiply function}
\usage{
atamult(A, Y, flag = 0)
}
\arguments{
\item{A}{a matrix}

\item{Y}{a vector}

\item{flag}{indicate with multiply of A and Y}
}
\value{
either A'*A*Y or A*Y or A'*Y
}
\description{
Jacobian matrix multiply function
}
\details{
compute either A'*A*Y or A*Y or A'*Y
}
\author{
Zhenglei Gao
}

