% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCS2.r
\docType{data}
\name{BCS2}
\alias{BCS2}
\title{BCS2}
\format{one parent and 5 metabolites, 6 column of residuals.}
\source{
BcS
}
\usage{
BCS1
}
\description{
This data set contains one parent and 5 metabolites.
}
\author{
2012-03-05
}

