% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.r
\docType{data}
\name{ex1}
\alias{ex1}
\title{ex1}
\format{list of kinetic model components}
\source{
BcS
}
\usage{
ex1
}
\description{
This data set contains the kinetic model with default starting values from data BCS1.
}
\author{
2013-03-05
}

