% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ex1_a}
\alias{ex1_a}
\title{ex1_a}
\format{list of kinetic model components}
\source{
BcS
}
\usage{
ex1_a
}
\description{
This data set contains the kinetic model with non-default starting values from data BCS1.
}
\author{
2013-03-05
}

