% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2.r
\docType{data}
\name{ex2}
\alias{ex2}
\title{ex2}
\format{list of kinetic model components}
\source{
BcS
}
\usage{
ex2
}
\description{
This data set contains the kinetic model with default starting values from data BCS2.
}
\author{
2013-03-05
}

