% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Fit_Schaefer}
\alias{Fit_Schaefer}
\title{Fit_Schaefer}
\format{list of kinetic model fit results}
\source{
BcS
}
\usage{
Fit_Schaefer
}
\description{
This data set contains fitting results from data Schaefer.
}
\author{
2013-03-05
}

