% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOCUS2006A.r
\docType{data}
\name{FOCUS2006A}
\alias{FOCUS2006A}
\title{Focus 2006 A}
\format{only parent substance instance, 1 column of residuals.}
\source{
Focus Kinetics
}
\usage{
FOCUS2006A
}
\description{
This data set contains only one parent residual data.
}
\author{
2012-03-05
}

