% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOCUS2006B.r
\docType{data}
\name{FOCUS2006B}
\alias{FOCUS2006B}
\title{Focus 2006 B}
\format{only parent substance instance, 1 column of residuals.}
\source{
Focus Kinetics
}
\usage{
FOCUS2006B
}
\description{
This data set contains only one parent residual data.
}
\author{
2012-03-05
}

