% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOCUS2006C.r
\docType{data}
\name{FOCUS2006C}
\alias{FOCUS2006C}
\title{Focus 2006 C}
\format{only parent substance instance, 1 column of residuals.}
\source{
Focus Kinetics
}
\usage{
FOCUS2006C
}
\description{
This data set contains only one parent residual data.
}
\author{
2012-03-05
}

