% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOCUS2006D.r
\docType{data}
\name{FOCUS2006D}
\alias{FOCUS2006D}
\title{Focus 2006 D}
\format{only parent substance instance, 1 column of residuals.}
\source{
Focus Kinetics
}
\usage{
FOCUS2006D
}
\description{
This data set contains only one parent residual data.
}
\author{
2012-03-05
}

