% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.r
\name{FOMC.solution}
\alias{FOMC.solution}
\title{a modification of the FOMC model solution function.}
\usage{
FOMC.solution(t, parent.0, alpha, beta)
}
\arguments{
\item{t}{time points}

\item{parent.0}{Concentration or state at time 0}

\item{alpha}{parameter for FOMC}

\item{beta}{parameter for FOMC}
}
\description{
Modified function based on \code{mkin::FOMC.solution}
}
\author{
Zhenglei Gao
}
\seealso{
\code{mkin::FOMC.solution}
}

