% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KineticEval-package.r
\docType{package}
\name{KineticEval-package}
\alias{KineticEval-package}
\title{The KineticEval Package}
\description{
Kinetic Evaluations by NLLS, IRLS and MCMC.
}
\details{
\tabular{ll}{
Package: \tab KineticEval\cr
Type: \tab Package\cr
Version: \tab 1.0-46\cr
Date: \tab 2016-10-07\cr
Depends: \tab R (>= 2.15), mkin,minqa\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://zhenglei-gao.github.io/KineticEval/\cr
}

This package is developed based on the R package `mkin`. It is
used by the KinGUI v2.1. The package is used to determine kinetic
parameters from results of environmental fate studies, e.g. aerobic soil
degradation, by fitting respective kinetic models to the observed data.The
package allows the choice between different optimization algorithms. In
particular the estimation of parameter confidence intervals is much
improved/corrected as compared to other commonly used evaluation softwares
by providing the methods Iteratively Reweighted Least Squares (IRLS) and
Markov Chain Monte Carlo (MCMC).
}
\author{
Zhenglei Gao
}
\keyword{package}

