\docType{package}
\name{KinGUI2-package}
\alias{KinGUI2-package}
\title{Kinetic Evaluations in R}
\description{
  In-line documentation for R.
}
\details{
  Roxygen is a Doxygen-like documentation system for R;
  allowing in-source specification of Rd files, collation
  and namespace directives.

  If you have existing Rd files, check out the
  \code{Rd2roxygen} package for a convenient way of
  converting Rd files to roxygen comments.

  \tabular{ll}{ Package: \tab KinGUI2\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab
  2012-03-15\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\note{
  This is based on the package \code{mkin} by Johannes
  Reinken.

  You are welcome to report other serious issues and bugs
  via \url{http://dropbox.com}.
}
\examples{
## see the package vignette: vignette('KinGUI2')
}
\author{
  Zhenglei Gao \email{zhenglei.gao@bayer.com}

  Maintainer: Zhenglei Gao \email{zhenglei.gao@bayer.com}
}
\keyword{package}

