% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list2ascii}
\alias{list2ascii}
\title{Write the results}
\usage{
list2ascii(x, file = paste(deparse(substitute(x)), ".txt", sep = ""))
}
\arguments{
\item{x}{a list.}

\item{file}{file name}
}
\description{
Function to write a list into a file.
}

