\name{norm}
\alias{norm}
\title{Compute the 2-norm of a vector}
\usage{
  norm(x)
}
\arguments{
  \item{x}{a vector}
}
\value{
  the 2-norm of the vector
}
\description{
  the 2-norm is the sum of the squares of the elements
}
\details{
  see also the R core function \code{norm}
}
\author{
  Zhenglei Gao
}

