% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinEval.r
\name{print.summary.kingui}
\alias{print.summary.kingui}
\title{Print method for \code{\link{summary.kingui}}}
\usage{
\method{print}{summary.kingui}(x, digits = max(3, getOption("digits") - 3),
  detailed = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{summary.kingui}}

\item{digits}{How many digits should be printed after the decimal point.}

\item{detailed}{Ignore for now.}

\item{...}{Other parameters to be passed into \code{\link{summary.kingui}}}
}
\value{
\code{NULL}
}
\description{
Print method for \code{\link{summary.kingui}}
}
\details{
Expanded from \code{\link{print.summary.modFit}} and \code{\link{print.summary.mkinfit}}
}
\author{
Zhenglei Gao
}

