% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmckinfit.full.r
\name{print.summary.mcmckingui}
\alias{print.summary.mcmckingui}
\title{S3 print method for class \code{summary.mcmckingui}}
\usage{
\method{print}{summary.mcmckingui}(x, digits = max(3, getOption("digits") -
  3), detailed = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mcmckingui}.}

\item{digits}{Number of digits to be printed after the decimal point.}

\item{detailed}{Not used.}

\item{...}{Optional arguments.}
}
\value{
\code{NULL}
}
\description{
S3 print method for class \code{summary.mcmckingui}
}
\author{
Zhenglei Gao
}

