% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrew.r
\docType{data}
\name{andrew}
\alias{andrew}
\title{Data from Andrew}
\format{including parent substance instance, 4 column of residuals.}
\source{
Focus Kinetics
}
\usage{
andrew
}
\description{
This data set contains multiple compartments.
}
\author{
2012-03-05
}

