% Generated by roxygen2 (4.0.2): do not edit by hand
\name{atBoundary}
\alias{atBoundary}
\title{check boundary reached}
\usage{
atBoundary(par, lower, upper)
}
\arguments{
\item{par}{fitted parameter vector}

\item{lower}{lower boundary}

\item{upper}{upper boundary}
}
\value{
TRUE or FALSE
}
\description{
check boundary reached
}
\author{
Zhenglei Gao
}

