% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmckinfit.full.r
\name{autolayout}
\alias{autolayout}
\title{Set the layout when making multiple plots on the same page.}
\usage{
autolayout(n)
}
\arguments{
\item{n}{number of plots to make}
}
\value{
number of rows and columns in the layout
}
\description{
Utility function to set the layout when making multiple plots on the same page.
}
\author{
Zhenglei Gao
}

