% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCS1.r
\docType{data}
\name{BCS1}
\alias{BCS1}
\title{BCS1}
\format{only parent and metab substance instance, 2 column of residuals.}
\source{
BcS
}
\usage{
BCS1
}
\description{
This data set contains only one parent residual data and one metabolite.
}
\author{
2012-03-05
}

