% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Schaefer.r
\docType{data}
\name{Fit_Schaefer}
\alias{Fit_Schaefer}
\title{Fit_Schaefer}
\format{list of kinetic model fit results}
\source{
BcS
}
\usage{
Fit_Schaefer
}
\description{
This data set contains fitting results from data Schaefer.
}
\author{
2013-03-05
}

