% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_mod.r
\name{kin_mod}
\alias{kin_mod}
\title{predicted residue values using defined kinetic model}
\usage{
kin_mod(P, inside = FALSE, plot = TRUE, plottitle, pnames0 = NULL, ...)
}
\arguments{
\item{P}{a vector of parameters}

\item{inside}{logical, whether the function is evaluated inside a fit function.}

\item{plot}{logical, whether to make a plot}

\item{plottitle}{the title of the plot}

\item{pnames0}{parameter names nls function will forget them!}

\item{...}{other parameters to be passed into plot}
}
\value{
the calculated model values
}
\description{
calculte the predicted/estimated residue values for kinetic models
}
\author{
Zhenglei Gao
}

