% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRR.r
\name{norm}
\alias{norm}
\title{Compute the 2-norm of a vector}
\usage{
norm(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical, whether to remove NAs.}
}
\value{
the 2-norm of the vector
}
\description{
the 2-norm is the sum of the squares of the elements
}
\details{
see also the R core function \code{norm}
}
\author{
Zhenglei Gao
}

