% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinEval.r
\name{rmse}
\alias{rmse}
\title{Calculating the root mean squared error.}
\usage{
rmse(obs, pred)
}
\arguments{
\item{obs}{Observed values.}

\item{pred}{Predicted values using model.}
}
\value{
The root mean squared error.
}
\description{
Calculating the scaled root mean squared error according to Focus Kinetics Guidelines.
}
\author{
Zhenglei Gao
}

