% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRR.r
\name{seceqn}
\alias{seceqn}
\title{Secular equation}
\usage{
seceqn(lambda, eigval, alpha, delta)
}
\arguments{
\item{lambda}{a vector}

\item{eigval}{eigenval}

\item{alpha}{alpha}

\item{delta}{delta}
}
\value{
the value of the secular equation at lambda
}
\description{
Secular equation
}
\details{
Secular equation is
}
\author{
Zhenglei Gao
}

