\name{summary.mkinmod.full}
\alias{summary.mkinmod.full}
\title{summary method for mkinmod.full}
\usage{
  summary.mkinmod.full(mkinmodini)
}
\arguments{
  \item{mkinmodini}{Kinetic Model or fit results with all
  parameters fixed}
}
\value{
  a list containing DT50 and other information
}
\description{
  Summary function for Kinetic models
}
\author{
  Zhenglei Gao
}

