% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.mkinmod.full}
\alias{summary.mkinmod.full}
\title{summary method for mkinmod.full}
\usage{
\method{summary}{mkinmod.full}(mkinmodini, ctr = kingui.control(),
  version = NULL)
}
\arguments{
\item{mkinmodini}{Kinetic Model or fit results with all parameters fixed}
}
\value{
a list containing DT50 and other information
}
\description{
Summary function for Kinetic models
}
\author{
Zhenglei Gao
}

