\docType{data}
\name{versioninfo}
\alias{versioninfo}
\title{Version information}
\format{1 column of version number, 1 column of bus and features, 1 column of status}
\source{
  Zhenglei Gao
}
\description{
  This data set contains version information for
  KineticEval
}
\author{
  Zhenglei Gao
}
\keyword{datasets}

