% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioninfo.r
\docType{data}
\name{versioninfo}
\alias{versioninfo}
\title{Version information}
\format{1 column of version number, 1 column of bus and features, 1 column of status}
\source{
Zhenglei Gao
}
\usage{
versioninfo
}
\description{
This data set contains version information for KineticEval
}
\author{
Zhenglei Gao
}

